// front-exam/src/app/apps/common/model/clientListDTO.ts
/**
 * Exam Management API
 * This API exposes endpoints to manage client.
 *
 * OpenAPI spec version: 1.0
 * Contact: yessinelleuch00@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ClientDTO } from './clientDTO';

export interface ClientListDTO { 
    clients?: Array<ClientDTO>;
    totalElements?: number;
    totalPages?: number;
}